Fix bug file delete before InputStream is closed
The catch clause would Files.delete the file p before the finally clause would close the InputStream in. This is problematic on Windows, because opening the file locks it, therefore it cannot be deleted. 
This fix uses try-with-resources from Java 7, which closes the resource introduced between the brackets before running the catch clause.